requirements.txt
streamlit==1.35.0
PyPDF2==3.0.1
langchain==0.2.1
langchain-google-genai==0.1.3
langchain-community==0.2.1  # Add this
faiss-cpu==1.8.0
python-dotenv==1.0.1
google-generativeai==0.5.2  # Add this
protobuf==4.25.3  # Add this